requirements.txt
dj-database-url==0.4.2
Django==2.0.2
djangorestframework==3.7.7
psycopg2==2.7.4
pytz==2018.3
waitress==1.1.0
whitenoise==3.3.1
